Expose Client and MailService classes, document use cases
