+++ packages/mail/src/mail.js
@@ -1,123 +1,9 @@
+'use strict';
 
 /**
  * Dependencies
  */
-const client = require('@sendgrid/client');
-const {
-  classes: {
-    Mail,
-  },
-} = require('@sendgrid/helpers');
+const MailService = require('./classes/mail-service');
 
-/**
- * Mail service class
- */
-class MailService {
-
-	/**
-	 * Constructor
-	 */
-  constructor() {
-    this.client = client;
-    this.substitutionWrappers = ['{{', '}}'];
-  }
-
-  /**
-   * API key pass through for convenience
-   */
-  setApiKey(apiKey) {
-    this.client.setApiKey(apiKey);
-  }
-
-  /**
-   * Set substitution wrappers
-   */
-  setSubstitutionWrappers(left, right) {
-    if (typeof left === 'undefined' || typeof right === 'undefined') {
-      throw new Error('Must provide both left and right side wrappers');
-    }
-    this.substitutionWrappers[0] = left;
-    this.substitutionWrappers[1] = right;
-  }
-
-  /**
-   * Send email
-   */
-  send(data, isMultiple = false, cb) {
-
-    //Callback as second parameter
-    if (typeof isMultiple === 'function') {
-      cb = isMultiple;
-      isMultiple = false;
-    }
-
-    //Array? Send in parallel
-    if (Array.isArray(data)) {
-
-      //Create promise
-      const promise = Promise.all(data.map(item => {
-        return this.send(item, isMultiple);
-      }));
-
-      //Execute callback if provided
-      if (cb) {
-        promise
-          .then(result => cb(null, result))
-          .catch(error => cb(error, null));
-      }
-
-      //Return promise
-      return promise;
-    }
-
-    //Send mail
-    try {
-
-      //Append multiple flag to data if not set
-      if (typeof data.isMultiple === 'undefined') {
-        data.isMultiple = isMultiple;
-      }
-
-      //Append global substitution wrappers if not set in data
-      if (typeof data.substitutionWrappers === 'undefined') {
-        data.substitutionWrappers = this.substitutionWrappers;
-      }
-
-      //Create Mail instance from data and get JSON body for request
-      const mail = Mail.create(data);
-      const body = mail.toJSON();
-
-      //Create request
-      const request = {
-        method: 'POST',
-        url: '/v3/mail/send',
-        body,
-      };
-
-      //Send
-      return this.client.request(request, cb);
-    }
-
-    //Catch sync errors
-    catch (error) {
-
-      //Pass to callback if provided
-      if (cb) {
-        cb(error, null);
-      }
-
-      //Reject promise
-      return Promise.reject(error);
-    }
-  }
-
-  /**
-   * Send multiple emails (shortcut)
-   */
-  sendMultiple(data, cb) {
-    return this.send(data, true, cb);
-  }
-}
-
-//Export class
+//Export singleton instance
 module.exports = new MailService();
